+++ src/Handler/PaymentOption/PaymentOptionHandler.php
@@ -37,11 +37,13 @@
 namespace Mollie\Handler\PaymentOption;
 
 use Configuration;
+use Customer;
 use Mollie\Api\Types\PaymentMethod;
 use Mollie\Config\Config;
 use Mollie\Provider\PaymentOption\BasePaymentOptionProvider;
 use Mollie\Provider\PaymentOption\CreditCardPaymentOptionProvider;
 use Mollie\Provider\PaymentOption\IdealPaymentOptionProvider;
+use Mollie\Repository\MolCustomerRepository;
 use MolPaymentMethod;
 
 class PaymentOptionHandler implements PaymentOptionHandlerInterface
@@ -60,15 +62,27 @@ class PaymentOptionHandler implements PaymentOptionHandlerInterface
      * @var IdealPaymentOptionProvider
      */
     private $idealPaymentOptionProvider;
+    /**
+     * @var MolCustomerRepository
+     */
+    private $customerRepository;
+    /**
+     * @var Customer
+     */
+    private $customer;
 
     public function __construct(
         BasePaymentOptionProvider $basePaymentOptionProvider,
         CreditCardPaymentOptionProvider $creditCardPaymentOptionProvider,
-        IdealPaymentOptionProvider $idealPaymentOptionProvider
+        IdealPaymentOptionProvider $idealPaymentOptionProvider,
+        MolCustomerRepository $customerRepository,
+        Customer $customer
     ) {
         $this->basePaymentOptionProvider = $basePaymentOptionProvider;
         $this->creditCardPaymentOptionProvider = $creditCardPaymentOptionProvider;
         $this->idealPaymentOptionProvider = $idealPaymentOptionProvider;
+        $this->customerRepository = $customerRepository;
+        $this->customer = $customer;
     }
 
     /**
@@ -113,9 +127,8 @@ class PaymentOptionHandler implements PaymentOptionHandlerInterface
     private function isCreditCardPaymentMethod(MolPaymentMethod $paymentMethod)
     {
         $isCreditCardPaymentMethod = PaymentMethod::CREDITCARD === $paymentMethod->getPaymentMethodName();
-        $isCartesBancairesPaymentMethod = Config::CARTES_BANCAIRES === $paymentMethod->getPaymentMethodName();
 
-        if (!$isCreditCardPaymentMethod && !$isCartesBancairesPaymentMethod) {
+        if (!$isCreditCardPaymentMethod) {
             return false;
         }