+++ changelog.md
@@ -14,6 +14,7 @@
 + Fixed issue where order product price would have wrong price if it didn't use default currency.
 + Added exception catch and added error message for Klarna status error when saving status when Klarna statuses are deleted.
 + Added check that doesn't allow enabling module if php version is not supported.
++ Added credit card single click logic with components.
 
 ## Changes in release 4.4.3 ##
 + Bank transfer now creates order on open status.